test: move the test_utils.dart file to the root test/ folder
