+++ packages/dart/test/src/objects/parse_object/parse_object_increment_decrement_test.dart
@@ -8,7 +8,7 @@ import 'package:test/test.dart';
 
 import '../../../parse_query_test.mocks.dart';
 import 'parse_object_test.dart';
-import 'test_utils.dart';
+import '../../../test_utils.dart';
 
 void main() {
   group('Increment/Decrement', () {
